Pundit Console is now working in VS
git-svn-id: https://pundit.svn.codeplex.com/svn/trunk@84823 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ src/Pundit.WinForms.Core/GlobalSettingsControl.cs
@@ -120,14 +120,17 @@ namespace Pundit.WinForms.Core
 
       private void cmdClearCache_Click(object sender, EventArgs e)
       {
-         foreach(string file in Directory.GetFiles(LocalRepository.GlobalRootFilePath))
+         if (DialogResult.Yes == Alert.AskYesNo(Strings.Alert_PurgeCache))
          {
-            try
+            foreach (string file in Directory.GetFiles(LocalRepository.GlobalRootFilePath))
             {
-               File.Delete(file);
-            }
-            catch
-            {
+               try
+               {
+                  File.Delete(file);
+               }
+               catch
+               {
+               }
             }
          }
       }
